requirements.txt.txt
requests
snowflake-connector-python